requirements.txt
apache-airflow==2.6.3
pylint==2.17.4
pyright==1.1.318
bandit==1.7.5
flake8==6.0.0
apache-airflow-providers-sqlite==3.4.2
requests==2.31.0
